Follow up fix, to resolve #8
In addition to setting the target SDK version in the manifest,
I needed to not write out the menu by returning false from the onCreateOptionsMenu
method call, which is used to initialize the contents of the Activity's standard
options menu. If I had things to put in there I would, but I don't so I can
override all of it to get rid of the 3 dot menu.